images and backend refactored

+++ backend/app/routes/categories.py
@@ -4,6 +4,7 @@ from app import db_helpers
 # from .utils import save_file, get_parent_ids  # Adjust import if needed
 from werkzeug.utils import secure_filename
 from app.utils import save_file, get_parent_ids, get_subcategory_ids  # Adjust import if needed
+from app.routes.images import delete_image_by_id
 
 categories_bp = Blueprint('category', __name__)
 
@@ -218,13 +219,31 @@ def delete_category():
     else:
         category_ids = get_subcategory_ids([category_id])
         # Delete the members
-        wildlife_ids_to_delete = [x["id"] for x in
-                                  db_helpers.select_multiple(f"SELECT id FROM Wildlife WHERE category_id IN ({','.join('?' for _ in category_ids)})", category_ids)]
-        db_helpers.delete(f"DELETE FROM Wildlife WHERE id IN ({','.join('?' for _ in wildlife_ids_to_delete)})",
-                          wildlife_ids_to_delete)
-        db_helpers.delete(
-            f"DELETE FROM FieldValues WHERE wildlife_id IN ({','.join('?' for _ in wildlife_ids_to_delete)})",
-            wildlife_ids_to_delete)
+        wildlife_ids_to_delete = [
+            x["id"] for x in db_helpers.select_multiple(
+                f"SELECT id FROM Wildlife WHERE category_id IN ({','.join('?' for _ in category_ids)})",
+                category_ids
+            )
+        ]
+
+        for wildlife_id in wildlife_ids_to_delete:
+            wildlife_images = db_helpers.select_multiple("SELECT id FROM Images WHERE wildlife_id = ?", [wildlife_id])
+            for image in wildlife_images:
+                delete_image_by_id(image["id"])
+
+        if wildlife_ids_to_delete:
+            db_helpers.delete(
+                f"DELETE FROM Wildlife WHERE id IN ({','.join('?' for _ in wildlife_ids_to_delete)})",
+                wildlife_ids_to_delete
+            )
+            db_helpers.delete(
+                f"DELETE FROM FieldValues WHERE wildlife_id IN ({','.join('?' for _ in wildlife_ids_to_delete)})",
+                wildlife_ids_to_delete
+            )
+
         # Delete the category and its subcategories
-        db_helpers.delete(f"DELETE FROM Categories WHERE id IN ({','.join('?' for _ in category_ids)})", category_ids)
+        db_helpers.delete(
+            f"DELETE FROM Categories WHERE id IN ({','.join('?' for _ in category_ids)})",
+            category_ids
+        )
         return jsonify({"message": "Category members and category successfully deleted"}), 200
